requirements.txt
alpaca-trade-api==3.0.0
yfinance==0.2.28
pandas==2.1.4
numpy==1.24.3
ta==0.10.2
twilio==8.10.0
google-generativeai==0.3.2
requests==2.31.0
python-dotenv==1.0.0
schedule==1.2.0
